Dev0.5.0

+++ fastNLP/io/loader/classification.py
@@ -1,19 +1,33 @@
-from ...core.dataset import DataSet
-from ...core.instance import Instance
-from .loader import Loader
-import warnings
+"""undocumented"""
+
+__all__ = [
+    "YelpLoader",
+    "YelpFullLoader",
+    "YelpPolarityLoader",
+    "IMDBLoader",
+    "SSTLoader",
+    "SST2Loader",
+    "ChnSentiCorpLoader"
+]
+
+import glob
 import os
 import random
 import shutil
-import numpy as np
+import time
+import warnings
+
+from .loader import Loader
+from ...core.dataset import DataSet
+from ...core.instance import Instance
+
 
 class YelpLoader(Loader):
     """
-    别名：:class:`fastNLP.io.YelpLoader` :class:`fastNLP.io.loader.YelpLoader`
-
     原始数据中内容应该为, 每一行为一个sample，第一个逗号之前为target，第一个逗号之后为文本内容。
 
     Example::
+    
         "1","I got 'new' tires from the..."
         "1","Don't waste your time..."
 
@@ -28,11 +42,11 @@ class YelpLoader(Loader):
        "...", "..."
 
     """
-
+    
     def __init__(self):
         super(YelpLoader, self).__init__()
-
-    def _load(self, path: str=None):
+    
+    def _load(self, path: str = None):
         ds = DataSet()
         with open(path, 'r', encoding='utf-8') as f:
             for line in f:
@@ -55,7 +69,7 @@ class YelpLoader(Loader):
 
 
 class YelpFullLoader(YelpLoader):
-    def download(self, dev_ratio: float = 0.1, seed: int = 0):
+    def download(self, dev_ratio: float = 0.1, re_download: bool = False):
         """
         自动下载数据集，如果你使用了这个数据集，请引用以下的文章
 
@@ -66,35 +80,23 @@ class YelpFullLoader(YelpLoader):
         dev.csv三个文件。
 
         :param float dev_ratio: 如果路径中没有dev集，从train划分多少作为dev的数据. 如果为0，则不划分dev。
-        :param int seed: 划分dev时的随机数种子
+        :param bool re_download: 是否重新下载数据，以重新切分数据。
         :return: str, 数据集的目录地址
         """
-
+        
         dataset_name = 'yelp-review-full'
         data_dir = self._get_dataset_path(dataset_name=dataset_name)
-        if os.path.exists(os.path.join(data_dir, 'dev.csv')):  # 存在dev的话，check是否需要重新下载
-            re_download = True
-            if dev_ratio>0:
-                dev_line_count = 0
-                tr_line_count = 0
-                with open(os.path.join(data_dir, 'train.csv'), 'r', encoding='utf-8') as f1, \
-                        open(os.path.join(data_dir, 'dev.csv'), 'r', encoding='utf-8') as f2:
-                    for line in f1:
-                        tr_line_count += 1
-                    for line in f2:
-                        dev_line_count += 1
-                if not np.isclose(dev_line_count, dev_ratio*(tr_line_count + dev_line_count), rtol=0.005):
-                    re_download = True
-                else:
-                    re_download = False
-            if re_download:
-                shutil.rmtree(data_dir)
-                data_dir = self._get_dataset_path(dataset_name=dataset_name)
-
+        modify_time = 0
+        for filepath in glob.glob(os.path.join(data_dir, '*')):
+            modify_time = os.stat(filepath).st_mtime
+            break
+        if time.time() - modify_time > 1 and re_download:  # 通过这种比较丑陋的方式判断一下文件是否是才下载的
+            shutil.rmtree(data_dir)
+            data_dir = self._get_dataset_path(dataset_name=dataset_name)
+        
         if not os.path.exists(os.path.join(data_dir, 'dev.csv')):
             if dev_ratio > 0:
                 assert 0 < dev_ratio < 1, "dev_ratio should be in range (0,1)."
-                random.seed(int(seed))
                 try:
                     with open(os.path.join(data_dir, 'train.csv'), 'r', encoding='utf-8') as f, \
                             open(os.path.join(data_dir, 'middle_file.csv'), 'w', encoding='utf-8') as f1, \
@@ -109,49 +111,37 @@ class YelpFullLoader(YelpLoader):
                 finally:
                     if os.path.exists(os.path.join(data_dir, 'middle_file.csv')):
                         os.remove(os.path.join(data_dir, 'middle_file.csv'))
-
+        
         return data_dir
 
 
 class YelpPolarityLoader(YelpLoader):
-    def download(self, dev_ratio: float = 0.1, seed: int = 0):
+    def download(self, dev_ratio: float = 0.1, re_download=False):
         """
         自动下载数据集，如果你使用了这个数据集，请引用以下的文章
 
         Xiang Zhang, Junbo Zhao, Yann LeCun. Character-level Convolutional Networks for Text Classification. Advances
         in Neural Information Processing Systems 28 (NIPS 2015)
 
-        根据dev_ratio的值随机将train中的数据取出一部分作为dev数据。下载完成后从train中切分0.1作为dev
+        根据dev_ratio的值随机将train中的数据取出一部分作为dev数据。下载完成后从train中切分dev_ratio这么多作为dev
 
-        :param float dev_ratio: 如果路径中不存在dev.csv, 从train划分多少作为dev的数据. 如果为0，则不划分dev
-        :param int seed: 划分dev时的随机数种子
+        :param float dev_ratio: 如果路径中不存在dev.csv, 从train划分多少作为dev的数据。 如果为0，则不划分dev。
+        :param bool re_download: 是否重新下载数据，以重新切分数据。
         :return: str, 数据集的目录地址
         """
         dataset_name = 'yelp-review-polarity'
         data_dir = self._get_dataset_path(dataset_name=dataset_name)
-        if os.path.exists(os.path.join(data_dir, 'dev.csv')):  # 存在dev的话，check是否符合比例要求
-            re_download = True
-            if dev_ratio>0:
-                dev_line_count = 0
-                tr_line_count = 0
-                with open(os.path.join(data_dir, 'train.csv'), 'r', encoding='utf-8') as f1, \
-                        open(os.path.join(data_dir, 'dev.csv'), 'r', encoding='utf-8') as f2:
-                    for line in f1:
-                        tr_line_count += 1
-                    for line in f2:
-                        dev_line_count += 1
-                if not np.isclose(dev_line_count, dev_ratio*(tr_line_count + dev_line_count), rtol=0.005):
-                    re_download = True
-                else:
-                    re_download = False
-            if re_download:
-                shutil.rmtree(data_dir)
-                data_dir = self._get_dataset_path(dataset_name=dataset_name)
-
+        modify_time = 0
+        for filepath in glob.glob(os.path.join(data_dir, '*')):
+            modify_time = os.stat(filepath).st_mtime
+            break
+        if time.time() - modify_time > 1 and re_download:  # 通过这种比较丑陋的方式判断一下文件是否是才下载的
+            shutil.rmtree(data_dir)
+            data_dir = self._get_dataset_path(dataset_name=dataset_name)
+        
         if not os.path.exists(os.path.join(data_dir, 'dev.csv')):
             if dev_ratio > 0:
                 assert 0 < dev_ratio < 1, "dev_ratio should be in range (0,1)."
-                random.seed(int(seed))
                 try:
                     with open(os.path.join(data_dir, 'train.csv'), 'r', encoding='utf-8') as f, \
                             open(os.path.join(data_dir, 'middle_file.csv'), 'w', encoding='utf-8') as f1, \
@@ -166,14 +156,12 @@ class YelpPolarityLoader(YelpLoader):
                 finally:
                     if os.path.exists(os.path.join(data_dir, 'middle_file.csv')):
                         os.remove(os.path.join(data_dir, 'middle_file.csv'))
-
+        
         return data_dir
 
 
 class IMDBLoader(Loader):
     """
-    别名：:class:`fastNLP.io.IMDBLoader` :class:`fastNLP.io.loader.IMDBLoader`
-
     IMDBLoader读取后的数据将具有以下两列内容: raw_words: str, 需要分类的文本; target: str, 文本的标签
     DataSet具备以下的结构:
 
@@ -185,10 +173,10 @@ class IMDBLoader(Loader):
        "...", "..."
 
     """
-
+    
     def __init__(self):
         super(IMDBLoader, self).__init__()
-
+    
     def _load(self, path: str):
         dataset = DataSet()
         with open(path, 'r', encoding="utf-8") as f:
@@ -201,13 +189,13 @@ class IMDBLoader(Loader):
                 words = parts[1]
                 if words:
                     dataset.append(Instance(raw_words=words, target=target))
-
+        
         if len(dataset) == 0:
             raise RuntimeError(f"{path} has no valid data.")
-
+        
         return dataset
-
-    def download(self, dev_ratio: float = 0.1, seed: int = 0):
+    
+    def download(self, dev_ratio: float = 0.1, re_download=False):
         """
         自动下载数据集，如果你使用了这个数据集，请引用以下的文章
 
@@ -216,34 +204,22 @@ class IMDBLoader(Loader):
         根据dev_ratio的值随机将train中的数据取出一部分作为dev数据。下载完成后从train中切分0.1作为dev
 
         :param float dev_ratio: 如果路径中没有dev.txt。从train划分多少作为dev的数据. 如果为0，则不划分dev
-        :param int seed: 划分dev时的随机数种子
+        :param bool re_download: 是否重新下载数据，以重新切分数据。
         :return: str, 数据集的目录地址
         """
         dataset_name = 'aclImdb'
         data_dir = self._get_dataset_path(dataset_name=dataset_name)
-        if os.path.exists(os.path.join(data_dir, 'dev.txt')):   # 存在dev的话，check是否符合比例要求
-            re_download = True
-            if dev_ratio>0:
-                dev_line_count = 0
-                tr_line_count = 0
-                with open(os.path.join(data_dir, 'train.txt'), 'r', encoding='utf-8') as f1, \
-                        open(os.path.join(data_dir, 'dev.txt'), 'r', encoding='utf-8') as f2:
-                    for line in f1:
-                        tr_line_count += 1
-                    for line in f2:
-                        dev_line_count += 1
-                if not np.isclose(dev_line_count, dev_ratio*(tr_line_count + dev_line_count), rtol=0.005):
-                    re_download = True
-                else:
-                    re_download = False
-            if re_download:
-                shutil.rmtree(data_dir)
-                data_dir = self._get_dataset_path(dataset_name=dataset_name)
-
+        modify_time = 0
+        for filepath in glob.glob(os.path.join(data_dir, '*')):
+            modify_time = os.stat(filepath).st_mtime
+            break
+        if time.time() - modify_time > 1 and re_download:  # 通过这种比较丑陋的方式判断一下文件是否是才下载的
+            shutil.rmtree(data_dir)
+            data_dir = self._get_dataset_path(dataset_name=dataset_name)
+        
         if not os.path.exists(os.path.join(data_dir, 'dev.csv')):
             if dev_ratio > 0:
                 assert 0 < dev_ratio < 1, "dev_ratio should be in range (0,1)."
-                random.seed(int(seed))
                 try:
                     with open(os.path.join(data_dir, 'train.txt'), 'r', encoding='utf-8') as f, \
                             open(os.path.join(data_dir, 'middle_file.txt'), 'w', encoding='utf-8') as f1, \
@@ -258,14 +234,12 @@ class IMDBLoader(Loader):
                 finally:
                     if os.path.exists(os.path.join(data_dir, 'middle_file.txt')):
                         os.remove(os.path.join(data_dir, 'middle_file.txt'))
-
+        
         return data_dir
 
 
 class SSTLoader(Loader):
     """
-    别名：:class:`fastNLP.io.SSTLoader` :class:`fastNLP.io.loader.SSTLoader`
-
     读取之后的DataSet具有以下的结构
 
     .. csv-table:: 下面是使用SSTLoader读取的DataSet所具备的field
@@ -278,10 +252,10 @@ class SSTLoader(Loader):
     raw_words列是str。
 
     """
-
+    
     def __init__(self):
         super().__init__()
-
+    
     def _load(self, path: str):
         """
         从path读取SST文件
@@ -296,7 +270,7 @@ class SSTLoader(Loader):
                 if line:
                     ds.append(Instance(raw_words=line))
         return ds
-
+    
     def download(self):
         """
         自动下载数据集，如果你使用了这个数据集，请引用以下的文章
@@ -323,10 +297,10 @@ class SST2Loader(Loader):
 
     test的DataSet没有target列。
     """
-
+    
     def __init__(self):
         super().__init__()
-
+    
     def _load(self, path: str):
         """
         从path读取SST2文件
@@ -335,7 +309,7 @@ class SST2Loader(Loader):
         :return: DataSet
         """
         ds = DataSet()
-
+        
         with open(path, 'r', encoding='utf-8') as f:
             f.readline()  # 跳过header
             if 'test' in os.path.split(path)[1]:
@@ -356,7 +330,7 @@ class SST2Loader(Loader):
                         if raw_words:
                             ds.append(Instance(raw_words=raw_words, target=target))
         return ds
-
+    
     def download(self):
         """
         自动下载数据集，如果你使用了该数据集，请引用以下的文章
@@ -366,4 +340,60 @@ class SST2Loader(Loader):
         :return:
         """
         output_dir = self._get_dataset_path(dataset_name='sst-2')
+        return output_dir
+
+
+class ChnSentiCorpLoader(Loader):
+    """
+    支持读取的数据的格式为，第一行为标题(具体内容会被忽略)，之后一行为一个sample，第一个制表符之前被认为是label，第
+    一个制表符及之后认为是句子
+
+    Example::
+
+        label	raw_chars
+        1	這間酒店環境和服務態度亦算不錯,但房間空間太小~~
+        1	<荐书> 推荐所有喜欢<红楼>的红迷们一定要收藏这本书,要知道...
+        0	商品的不足暂时还没发现，京东的订单处理速度实在.......周二就打包完成，周五才发货...
+
+    读取后的DataSet具有以下的field
+
+    .. csv-table::
+        :header: "raw_chars", "target"
+
+        "這間酒店環境和服務態度亦算不錯,但房間空間太小~~", "1"
+        "<荐书> 推荐所有喜欢<红楼>...", "1"
+        "..."
+
+    """
+    def __init__(self):
+        super().__init__()
+
+    def _load(self, path:str):
+        """
+        从path中读取数据
+
+        :param path:
+        :return:
+        """
+        ds = DataSet()
+        with open(path, 'r', encoding='utf-8') as f:
+            f.readline()
+            for line in f:
+                line = line.strip()
+                tab_index = line.index('\t')
+                if tab_index!=-1:
+                    target = line[:tab_index]
+                    raw_chars = line[tab_index+1:]
+                    if raw_chars:
+                        ds.append(Instance(raw_chars=raw_chars, target=target))
+        return ds
+
+    def download(self)->str:
+        """
+        自动下载数据，该数据取自https://github.com/pengming617/bert_classification/tree/master/data，在
+        https://arxiv.org/pdf/1904.09223.pdf与https://arxiv.org/pdf/1906.08101.pdf有使用
+
+        :return:
+        """
+        output_dir = self._get_dataset_path('chn-senti-corp')
         return output_dir
